+++ Modules~/CountdownServiceModule/Scripts/Runtime/RootsContexts/CountdownServiceRoot.cs
@@ -1,9 +1,0 @@
-using FlowIoC.BaseModule.Root;
-
-namespace Modules.CountdownServiceModule.RootsContexts
-{
-    public class CountdownServiceRoot : Root<CountdownServiceContext>
-    {
-        
-    }
-}