+++ Modules~/CounterModule/Scripts/Runtime/RootsContexts/CounterServiceRoot.cs
@@ -1,0 +1,9 @@
+using FlowIoC.BaseModule.Root;
+
+namespace Modules.CounterModule.RootsContexts
+{
+    public class CounterServiceRoot : Root<CounterServiceContext>
+    {
+        
+    }
+}