+++ Modules~/CounterModule/zTestModules/CounterTestModule/Scripts/Runtime/RootsContexts/CounterTestRoot.cs
@@ -1,0 +1,11 @@
+#if UNITY_EDITOR
+using FlowIoC.BaseModule.Root;
+
+namespace Modules.CounterModule.CounterTestModule.RootsContexts
+{
+    public class CounterTestRoot : Root<CounterTestContext>
+    {
+        
+    }
+}
+#endif